+++ packages/components/src/components/Search/Search.tsx
@@ -2,14 +2,15 @@ import { MaterialIcons } from "@expo/vector-icons";
 import React, { useEffect, useRef, useState } from "react";
 import { Animated, View } from "react-native";
 import { TextInput } from "react-native-gesture-handler";
-import { Button, Input, TextFieldProps, useStyles, useToken } from "../..";
+import { Button, Input, InputProps, useStyles, useToken } from "../..";
 import { EDSStyleSheet } from "../../styling";
 
-export type SearchProps = Omit<TextFieldProps, "multiline"> & {
+export type SearchProps = Omit<InputProps, "multiline"> & {
     cancellable?: boolean;
     onCancelPress?: () => void;
+    textValue?: string;
 };
-export const Search = ({ cancellable, onCancelPress, ...restProps }: SearchProps) => {
+export const Search = ({ cancellable, onCancelPress, onChange, textValue, ...restProps }: SearchProps) => {
     const [text, setText] = useState('');
     const [isInputFocused, setIsInputFocused] = useState(false);
     const [cancelButtonWidth, setCancelButtonWidth] = useState<number>(0);
@@ -40,8 +41,13 @@ export const Search = ({ cancellable, onCancelPress, ...restProps }: SearchProps
             duration: token.timing.animation.slow,
             useNativeDriver: false
         }).start();
-    }, [isInputFocused]);
+    }, [animationValue, cancellable, isInputFocused, token.timing.animation.slow]);
 
+    useEffect(() => {
+        if (textValue) {
+            setText(textValue)
+        }
+    }, [textValue])
 
     const handleCancel = () => {
         setText('');
@@ -52,7 +58,13 @@ export const Search = ({ cancellable, onCancelPress, ...restProps }: SearchProps
     // FIXME: On web the search field loses focus when the clear text button is pressed. 
     const handleClearText = () => {
         setText('');
+        onChange?.('');
         inputRef.current?.focus();
+    }
+
+    const onChangeText = (text: string) => {
+        setText(text);
+        onChange?.(text);
     }
 
     return (
@@ -62,9 +74,15 @@ export const Search = ({ cancellable, onCancelPress, ...restProps }: SearchProps
                     {...restProps}
                     ref={inputRef}
                     value={text}
-                    onChange={setText}
-                    onFocus={() => setIsInputFocused(true)}
-                    onBlur={() => setIsInputFocused(false)}
+                    onChange={onChangeText}
+                    onFocus={(e) => {
+                        setIsInputFocused(true);
+                        restProps.onFocus?.(e);
+                    }}
+                    onBlur={(e) => {
+                        setIsInputFocused(false)
+                        restProps.onBlur?.(e);
+                    }}
                     leftAdornments={
                         <View style={styles.adornment}>
                             <MaterialIcons name="search" size={18} color={styles.icon.color} />
@@ -103,6 +121,7 @@ export const Search = ({ cancellable, onCancelPress, ...restProps }: SearchProps
 
 const themedStyles = EDSStyleSheet.create(theme => ({
     container: {
+        flex: 1,
         flexDirection: "row",
         alignItems: "flex-end",
     },