+++ packages/components/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @equinor/mad-components
 
+## 0.6.6
+
+### Patch Changes
+
+-   bd5ded0: Make search get its text value from other sources
+
+## 0.6.5
+
+### Patch Changes
+
+-   adaa7bb: fixed clear text button onchange issues on search
+
+## 0.6.4
+
+### Patch Changes
+
+-   85b85ea: fixed flex layout
+
+## 0.6.3
+
+### Patch Changes
+
+-   aba4173: fixed onchange and onblur issue
+
+## 0.6.2
+
+### Patch Changes
+
+-   1499432: fixed input onchange issue
+
 ## 0.6.1
 
 ### Patch Changes